requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
aiohttp==3.11.0
openai==1.54.0
google-generativeai==0.3.2
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
python-dotenv==1.0.1
pydantic==2.9.0
pydantic-settings==2.5.0
Pillow==10.2.0

